requirements_eval.txt
openai-whisper>=20231117
jiwer>=3.0.3
groq>=0.9.0
torch>=2.0.0
torchaudio>=2.0.0
ffmpeg-python>=0.2.0
numpy>=1.24.0
python-dotenv>=1.0.0